Change to f32, fix an infinite loop

+++ src/camera.rs
@@ -9,7 +9,7 @@ pub struct Camera {
     u: Vec3,
     v: Vec3,
     _w: Vec3,
-    lens_radius: f64,
+    lens_radius: f32,
 }
 
 impl Camera {
@@ -21,13 +21,13 @@ impl Camera {
         // Up reference vector.
         v_up: Vec3,
         // Vertical
-        v_fov: f64,
+        v_fov: f32,
         // Aspect ratio.
-        aspect: f64,
+        aspect: f32,
         // Set aperture to 0 to disable depth of field.
-        aperture: f64,
+        aperture: f32,
         // Focus distance.
-        focus_dist: f64,
+        focus_dist: f32,
     ) -> Self {
         let theta = v_fov.to_radians();
         let half_height = (theta / 2.0).tan();
@@ -50,7 +50,7 @@ impl Camera {
         }
     }
 
-    pub fn get_ray(&self, s: f64, t: f64) -> Ray {
+    pub fn get_ray(&self, s: f32, t: f32) -> Ray {
         let rd = Self::random_in_unit_disk() * self.lens_radius;
         let offset = self.u * rd.x + self.v * rd.y;
         Ray::new(
@@ -60,8 +60,8 @@ impl Camera {
     }
 
     fn random_in_unit_disk() -> Vec3 {
-        let theta = rand::random::<f64>() * 2.0 * std::f64::consts::PI;
-        let r = rand::random::<f64>().sqrt();
+        let theta = rand::random::<f32>() * 2.0 * std::f32::consts::PI;
+        let r = rand::random::<f32>().sqrt();
         vec3(r * theta.cos(), r * theta.sin(), 0.0)
     }
 }
